pyproject.toml
[project]
name = "exmat-ai"
version = "0.1.0"
description = "Automated Battery Material Data Extraction using Agentic AI"
authors = [
    {name = "Naga Sai Teja Kolakaleti", email = "Naga.Kolakaleti@uni-bayreuth.de"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers==4.37.2",
    
    # LangGraph
    "langgraph>=0.0.40",
    "langgraph-checkpoint>=0.0.6",
    
    # Object Detection
    "ultralytics>=8.0.0",
    
    # Chemistry
    "rdkit>=2023.9.1",
    "pubchempy>=1.0.4",
    
    # ML utilities
    "huggingface-hub>=0.19.4",
    "safetensors>=0.4.1",
    
    # Image/Document processing
    "Pillow>=10.1.0",
    "opencv-python>=4.8.1",
    "pdf2image>=1.16.3",
    
    # Data handling
    "pandas>=2.1.3",
    "numpy>=1.24.3",
    "openpyxl>=3.1.2",
    
    # Utilities
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.1",
    "colorama>=0.4.6",
    "loguru>=0.7.2",
    
    # Testing
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ExMat_AI"
Repository = "https://github.com/yourusername/ExMat_AI"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# CRITICAL: Define workspace explicitly to exclude DeepSeek-OCR
[tool.uv.workspace]
members = []  # No workspace members - this is a standalone project

[dependency-groups]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100
